Move eval_double tests to eval folder
